Components and input for Scripting
- Changed how WindowsWindow handles messages
- Changed Key Codes
- Changed MouseCodes to MouseButton and changed some button names
- TimeStep now uses float instead of double
- Entity UUID map for fast access to entities with uuids
- Added IsEditorScene flag to Scene
-Physics Debug Panel
Scripting:
- Added ScriptComponent
- Renamed MonoCalls to MonoGlue
C#
- Added Components
- Added Input

+++ Shark/src/Platform/DirectX11/DirectXGPUTimer.h
@@ -12,7 +12,7 @@ namespace Shark {
 		DirectXGPUTimer(const std::string& name);
 		virtual ~DirectXGPUTimer();
 
-		virtual TimeStep GetTime() const override { return (double)m_LastTickCount / (double)m_LastFrequency; }
+		virtual TimeStep GetTime() const override { return (float)m_LastTickCount / (float)m_LastFrequency; }
 
 		uint64_t GetTickCount() const { return m_LastTickCount; }
 		uint64_t GetFrequency() const { return m_LastFrequency; }
